pyproject.toml
[project]
name = "serpentone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "mido>=1.3.3",
    "pynput>=1.8.1",
    "python-rtmidi>=1.5.8",
    "supriya>=25.9b1",
    "textual>=6.6.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ty>=0.0.7",
]
